refactor<function>: Created a new file and move the functions from chat page to mediaPermission.ts and added new color palletes to UserPermission
