+++ client/src/components/media_permission/MediaComponent.tsx
@@ -17,8 +17,8 @@ export const MediaToggle: React.FC<MediaToggleProps> = ({
     <div className="flex items-center">
       {icon}
       <div>
-        <h4 className="text-lg font-medium text-gray-900">{title}</h4>
-        <p className="text-sm text-gray-500">{description}</p>
+        <h4 className="text-lg font-medium text-primaryPink">{title}</h4>
+        <p className="text-sm text-accentOrange">{description}</p>
       </div>
     </div>
     <div className="flex items-center">
@@ -29,7 +29,9 @@ export const MediaToggle: React.FC<MediaToggleProps> = ({
           checked={isPermissionGranted === true}
           onChange={onPermissionRequest}
         />
-        <div className="relative w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-indigo-300 rounded-full peer peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:bg-indigo-600"></div>
+
+        {/* Switch Button */}
+        <div className="relative w-11 h-6 bg-dangerRed peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-highlightOrange rounded-full peer peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:bg-primaryPink"></div>
       </label>
     </div>
   </div>
