+++ client/src/pages/ChatPage.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 
 import { Room } from "./Room";
 
-import UserPermission from "../components/UserPermission";
+import UserPermission from "../components/media_permission/UserPermission";
 
 import getBrowserType from "../utils/getBrowser";
